[FIX] fixed a too short cigar count field, made read_sam/bam more robust against accidental misuse

+++ core/include/seqan/bam_io/cigar.h
@@ -122,7 +122,7 @@ namespace seqan {
 ..type:nolink:$TCount$
 */
 
-template <typename TOperation_ = char, typename TCount_ = unsigned short>
+template <typename TOperation_ = char, typename TCount_ = unsigned>
 struct CigarElement
 {
     typedef TOperation_ TOperation;
